app/requirements.txt
streamlit>=1.30.0
pandas>=1.5.0
openpyxl>=3.0.0
pyyaml>=6.0
